// ui/src/app/core/api-client/model/server-instance-info-dto.ts
/**
 * API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ServerDescriptionDto } from './server-description-dto';
import { ServerPopulationDto } from './server-population-dto';
import { ServerVersionDto } from './server-version-dto';


export interface ServerInstanceInfoDto { 
    state?: ServerInstanceInfoDto.StateEnum;
    version?: ServerVersionDto;
    population?: ServerPopulationDto;
    description?: ServerDescriptionDto;
    icon?: string;
}
export namespace ServerInstanceInfoDto {
    export const StateEnum = {
        Stopped: 'STOPPED',
        Starting: 'STARTING',
        Running: 'RUNNING',
        Shutdown: 'SHUTDOWN'
    } as const;
    export type StateEnum = typeof StateEnum[keyof typeof StateEnum];
}